requirements.txt
Django==3.1.1
django-crispy-forms==1.8.1
django-rest-framework==0.1.0
mysqlclient==1.4.6
humanize==0.5.1
Pillow==6.2.1

